feat(`@vtmn/css`, `@vtmn/svelte`, `@vtmn/react`, `@vtmn/vue`): `VtmnChip` variant `input` match now the figma specs (#1456)
BREAKING CHANGE: The design of `VtmnChip`now matches Figma's version. If you used the `VtmnChip` with an "input" variant, you will have changes when you will upgrade to this new version. "Selected" input is no longer available on this variant. The color of the variant stays white, and no longer has a blue color.